packages/ai-personalization/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.4
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.3
deprecation==2.1.0
Flask==3.0.0
Flask-Cors==4.0.0
frozenlist==1.8.0
fsspec==2026.1.0
google-ai-generativelanguage==0.4.0
google-api-core==2.25.2
google-auth==2.47.0
google-generativeai==0.3.2
googleapis-common-protos==1.72.0
gotrue==1.3.1
grpcio==1.76.0
grpcio-status==1.62.3
gunicorn==21.2.0
h11==0.14.0
h2==4.3.0
hpack==4.1.0
httpcore==0.17.3
httpx==0.24.1
hyperframe==6.1.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mmh3==5.2.0
multidict==6.7.0
numpy==1.26.3
packaging==25.0
pandas==2.3.3
postgrest==0.15.1
propcache==0.4.1
proto-plus==1.27.0
protobuf==4.25.8
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyiceberg==0.10.0
PyJWT==2.10.1
pyparsing==3.3.1
pyroaring==1.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
realtime==1.0.6
requests==2.32.5
rich==14.2.0
rsa==4.9.1
scikit-learn==1.4.0
scipy==1.16.3
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
storage3==0.7.7
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.3.4
supabase-auth==2.27.2
supabase-functions==2.27.2
supafunc==0.3.3
tenacity==9.1.2
threadpoolctl==3.6.0
tqdm==4.67.1
typing_extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
urllib3==2.6.3
websockets==12.0
Werkzeug==3.1.5
yarl==1.22.0
